Change replace task, almost final version

+++ project/FilterResPlugin.scala
@@ -1,5 +1,5 @@
 import sbt.PluginTrigger.AllRequirements
-import sbt.{AutoPlugin, PluginTrigger, settingKey, taskKey}
+import sbt.{AutoPlugin, IO, PluginTrigger, settingKey, taskKey}
 
 import java.io.{File, PrintWriter}
 import scala.io.Source
@@ -10,30 +10,45 @@ object FilterResPlugin extends AutoPlugin {
   sealed trait Keys {
     lazy val replace = taskKey[Unit]("Task to replace variables from dictionary")
     lazy val dictionary = settingKey[Map[String, String]]("Variables dictionary")
+    lazy val extensions = settingKey[Seq[String]]("Config extensions")
+    lazy val mainVersion = settingKey[String]("Scala version")
+    lazy val from = settingKey[File]("Resource directory")
+    lazy val to = settingKey[File]("Artifact directory")
   }
 
   object autoImport extends Keys {
-    def replaceVars(res: File, dictionary: Map[String, String]): Unit = {
-      res.listFiles foreach {
-        case file if file.isFile => editFile(file, dictionary)
-        case file if file.isDirectory => replaceVars(file, dictionary)
+    def replaceVars(res: File, dictionary: Map[String, String], extensions: Seq[String]): Unit =
+      Option(res.listFiles) foreach {
+        _.foreach {
+          case file if file.isFile => editFile(file, dictionary, extensions)
+          case file if file.isDirectory => replaceVars(file, dictionary, extensions)
+          case _ =>
+        }
+      }
+
+    private def editFile(file: File, dictionary: Map[String, String], extensions: Seq[String]) = {
+      if (extensions.exists(file.getName.contains)) {
+        val tmp = new File(file.getPath.replace(file.getName, "tmp.conf"))
+        val out = new PrintWriter(tmp)
+        IO.readLines(file).map { line =>
+          dictionary.foldLeft(line) { case (cur, (from, to)) =>
+            cur.replace("${" + from + "}", to)
+          }
+        }.foreach(out.println)
+        out.close()
+        file.delete()
+        tmp.renameTo(file)
       }
     }
 
-    private def editFile(file: File, dictionary: Map[String, String]) = {
-      val in = Source.fromFile(file)
-      val tmp = new File(file.getPath.replace(file.getName, "tmp.conf"))
-      val out = new PrintWriter(tmp)
-      in.getLines.map { line =>
-        dictionary.foldLeft(line) { case (cur, (from, to)) =>
-          cur.replace("${" + from + "}", to)
-        }
-      }.foreach(out.println)
-      in.close()
-      out.close()
-      file.delete()
-      tmp.renameTo(file)
-    }
+/*    private def revertFile(file: File) = {
+      val extension = ".backup"
+      if (file.getName.contains(extension)) {
+        val tmp = new File(file.getPath.replace(extension, ""))
+        tmp.delete()
+        file.renameTo(tmp)
+      }
+    }*/
   }
 }
